refactor: make slider more usuable
