+++ src/features/user/MFV2/InfoAndCTA/InfoAndCta.module.scss
@@ -33,7 +33,7 @@
     }
 
     @media screen and (max-width: 460px) {
-        height: 230px;
+        height: 260px;
         width: 100%;
     }
 }
